Add functionality to preview cobradoc changes
This occurs on PRs opened against either the `main` branch or `release-`
branches. The strategy here is to always have the PR consist of two
commits; the first commit is the docs generated against the base ref of
the PR, and the second commit is the docs generated against the head ref
of the PR. It's possible, but unlikely, that the first commit will be
empty, but the second commit will always contain a diff. These preview PRs
should never be merged, anyway.

Signed-off-by: Andrew Mason <andrew@planetscale.com>