Run `find` in a cross-platform manner
Managed to finally get my hands on a relevant log entry and saw:

```
{"level":"error","github_event_type":"pull_request","github_delivery_id":"<omitted>","error":"Failed to `find` latest docs version to generate cobradocs preview for https://github.com/vitessio/vitess/pull/14789: exit status 1\nstderr: find: unknown predicate `-sE'\n\nstdout: ","time":"2024-01-24T13:13:22Z","message":"Unexpected error handling webhook"}
```

So, we need to remove `-sE` on unix, but not on darwin.

`-E` _before_ the path is replaced by `-regextype posix-extended` _after_
the path.

There is no replacement for `-s` on non-darwin unix, so we now need to
pipe to `sort` to sort in dictionary order.

Signed-off-by: Andrew Mason <andrew@planetscale.com>