Fetch by PR number rather than ref
This is how `cloneVitessAndGenerateErrors` clones; the advantage here is
when the PR is based on a branch of a remote _other_ than the repo the PR
is against (e.g. `planetscale/vitess` vs `vitessio/vitess`), we can fetch
the PR's `HEAD` without needing to add a remote for that repository.

Didn't come up in my local testing as I was opening PRs against the same
repo (`ajm188/vitess`).

Signed-off-by: Andrew Mason <andrew@planetscale.com>